I need to improve the algorithm I guess...
OrbProfile is so challenging for the AI. This is hard. Maybe I should just tell the AI how I do it? It doesn't really work as I expect. I can do it within 50 so the AI should able to do it as well. I guess combo is more important here. How should I change it?